🧹 Moved NEW_NOTE mutation from new.jsx
