+++ apps/react-tsn/src/gql/mutation.jsx
@@ -12,6 +12,26 @@ mutation signIn($email: String, $password: String!) {
       }
 `;
 
+// our new note query
+const NEW_NOTE = gql`
+mutation newNote($content: String!) {
+        newNote(content: $content) {
+          id
+          content
+          createdAt
+          favoriteCount
+          favoritedBy {
+id
+username 
+}
+author {
+username
+id 
+}
+} 
+}
+`;
+
 const EDIT_NOTE = gql`
 mutation updateNote($id: ID!, $content: String!) {
         updateNote(id: $id, content: $content) {
@@ -45,4 +65,11 @@ const TOGGLE_FAVORITE = gql` mutation toggleFavorite($id: ID!) {
 } 
 `;
 
-export { SIGNUP_USER, SIGNIN_USER, EDIT_NOTE, DELETE_NOTE, TOGGLE_FAVORITE };
+export {
+	SIGNUP_USER,
+	SIGNIN_USER,
+	NEW_NOTE,
+	EDIT_NOTE,
+	DELETE_NOTE,
+	TOGGLE_FAVORITE,
+};
