+++ apps/react-tsn/src/gql/query.js
@@ -18,7 +18,8 @@ query noteFeed($cursor: String) {
 } }
 `;
 
-const GET_NOTE = gql` query note($id: ID!) {
+const GET_NOTE = gql` 
+query note($id: ID!) {
     note(id: $id) {
       id
       createdAt
@@ -34,7 +35,8 @@ const GET_NOTE = gql` query note($id: ID!) {
 `;
 
 // add the GET_MY_NOTES query
-const GET_MY_NOTES = gql` query me {
+const GET_MY_NOTES = gql` 
+query me {
 me { id
           username
           notes {
@@ -53,7 +55,8 @@ me { id
 `;
 
 // add the GET_MY_FAVORITES query
-const GET_MY_FAVORITES = gql` query me {
+const GET_MY_FAVORITES = gql` 
+query me {
 me { 
 id
           username
