+++ apps/react-tsn/src/pages/new.jsx
@@ -1,33 +1,15 @@
 import React, { useEffect } from "react";
-import { useMutation, gql } from "@apollo/client";
+import { useMutation } from "@apollo/client";
 import { useNavigate } from "react-router-dom";
 // import the NoteForm component
 import NoteForm from "../components/NoteForm";
 
 import { GET_MY_NOTES, GET_NOTES } from "../gql/query";
+import { NEW_NOTE } from "../gql/mutation";
 
 // within the NewNote component update the mutation
 //everything else stays the same
 
-// our new note query
-const NEW_NOTE = gql`
-mutation newNote($content: String!) {
-        newNote(content: $content) {
-          id
-          content
-          createdAt
-          favoriteCount
-          favoritedBy {
-id
-username 
-}
-author {
-username
-id 
-}
-} 
-}
-`;
 const NewNote = (props) => {
 	useEffect(() => {
 		//update the document title
@@ -40,8 +22,7 @@ const NewNote = (props) => {
 		refetchQueries: [{ query: GET_MY_NOTES }, { query: GET_NOTES }],
 		onCompleted: (data) => {
 			// when complete, redirect the user to the note page
-			console.log(data);
-			// props.history.push(`note/${data.newNote.id}`);
+			console.log(data.newNote.createdAt);
 			navigate(`note/${data.newNote.id}`);
 		},
 	});
